version bump 0.5a2 removed markov modelling significant change to unobs estimation with severely overlapped pairs explicit logic for small samples users can control number of bootstraps
